# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GradeWise GH API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface AuthRequest {
  username: string;
}

export interface UserProfile {
  id: number;
  username: string;
  universityId?: number | null;
  universityName?: string | null;
  currentLevel?: number | null;
  themeColor?: string | null;
  customPrimaryColor?: string | null;
  customSecondaryColor?: string | null;
}

export interface AuthResponse {
  user: UserProfile;
}

export type UniversityGradingSystem =
  (typeof UniversityGradingSystem)[keyof typeof UniversityGradingSystem];

export const UniversityGradingSystem = {
  KNUST: "KNUST",
  UG: "UG",
  UCC: "UCC",
  UDS: "UDS",
  UMAT: "UMAT",
  GIMPA: "GIMPA",
  AUCC: "AUCC",
  STANDARD: "STANDARD",
} as const;

export interface GradePoint {
  grade: string;
  minScore: number;
  maxScore: number;
  points: number;
  interpretation: string;
}

export interface University {
  id: number;
  name: string;
  shortName: string;
  gradingSystem: UniversityGradingSystem;
  gradingScale: GradePoint[];
  firstClass: number;
  secondClassUpper: number;
  secondClassLower: number;
  thirdClass: number;
  pass: number;
}

export interface Course {
  id: number;
  name: string;
  creditHours: number;
  grade?: string | null;
  score?: number | null;
  gradePoints?: number | null;
}

export interface Semester {
  id: number;
  userId: number;
  level: number;
  semesterNumber: number;
  courses: Course[];
  gpa?: number | null;
}

export interface SaveCourseRequest {
  name: string;
  creditHours: number;
  grade?: string | null;
  score?: number | null;
}

export interface SaveSemesterRequest {
  level: number;
  semesterNumber: number;
  courses: SaveCourseRequest[];
}

export type CgpaGoalLikelihood =
  | (typeof CgpaGoalLikelihood)[keyof typeof CgpaGoalLikelihood]
  | null;

export const CgpaGoalLikelihood = {
  very_high: "very_high",
  high: "high",
  moderate: "moderate",
  low: "low",
  very_low: "very_low",
} as const;

export interface CgpaGoal {
  targetCgpa?: number | null;
  remainingCredits?: number | null;
  requiredGpa?: number | null;
  likelihood?: CgpaGoalLikelihood;
  currentCgpa?: number | null;
  classification?: string | null;
}

export interface SaveCgpaGoalRequest {
  targetCgpa: number;
  remainingCredits: number;
}

export interface SaveProfileRequest {
  universityId?: number | null;
  currentLevel?: number | null;
  themeColor?: string | null;
  customPrimaryColor?: string | null;
  customSecondaryColor?: string | null;
}