# lib/api-zod/src/generated/types/cgpaGoal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GradeWise GH API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CgpaGoalLikelihood } from "./cgpaGoalLikelihood";

export interface CgpaGoal {
  targetCgpa?: number | null;
  remainingCredits?: number | null;
  requiredGpa?: number | null;
  likelihood?: CgpaGoalLikelihood;
  currentCgpa?: number | null;
  classification?: string | null;
}