# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GradeWise GH API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Course {
  id: number;
  name: string;
  creditHours: number;
  grade?: string | null;
  score?: number | null;
  gradePoints?: number | null;
}