# lib/api-zod/src/generated/types/gradePoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GradeWise GH API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GradePoint {
  grade: string;
  minScore: number;
  maxScore: number;
  points: number;
  interpretation: string;
}