# lib/api-zod/src/generated/types/saveCourseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GradeWise GH API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SaveCourseRequest {
  name: string;
  creditHours: number;
  grade?: string | null;
  score?: number | null;
}