# lib/api-zod/src/generated/types/saveProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GradeWise GH API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SaveProfileRequest {
  universityId?: number | null;
  currentLevel?: number | null;
  themeColor?: string | null;
  customPrimaryColor?: string | null;
  customSecondaryColor?: string | null;
}