# lib/api-zod/src/generated/types/saveSemesterRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GradeWise GH API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SaveCourseRequest } from "./saveCourseRequest";

export interface SaveSemesterRequest {
  level: number;
  semesterNumber: number;
  courses: SaveCourseRequest[];
}