# lib/api-zod/src/generated/types/semester.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GradeWise GH API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Course } from "./course";

export interface Semester {
  id: number;
  userId: number;
  level: number;
  semesterNumber: number;
  courses: Course[];
  gpa?: number | null;
}