# lib/api-zod/src/generated/types/university.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GradeWise GH API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GradePoint } from "./gradePoint";
import type { UniversityGradingSystem } from "./universityGradingSystem";

export interface University {
  id: number;
  name: string;
  shortName: string;
  gradingSystem: UniversityGradingSystem;
  gradingScale: GradePoint[];
  firstClass: number;
  secondClassUpper: number;
  secondClassLower: number;
  thirdClass: number;
  pass: number;
}